# src/swagger/swagger.json
{
	"basePath": "/service",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"UserAttributes": {
			"properties": {
				"email": {
					"type": "string",
					"format": "email",
					"pattern": "^[a-zA-Z0-9_.+-]+\\x40[a-zA-Z0-9-]+\\.[a-zA-Z0-9-.]+$"
				},
				"name": {
					"type": "string",
					"minLength": 1
				},
				"role": {
					"type": "string"
				},
				"password": {
					"type": "string",
					"x-nullable": true
				},
				"id": {
					"type": "number",
					"format": "double",
					"x-nullable": true
				},
				"createdAt": {
					"type": "string",
					"format": "date-time",
					"x-nullable": true
				},
				"updatedAt": {
					"type": "string",
					"format": "date-time",
					"x-nullable": true
				}
			},
			"required": [
				"email",
				"name",
				"role"
			],
			"type": "object"
		},
		"PaginationResponseUserAttributes[]": {
			"properties": {
				"count": {
					"type": "number",
					"format": "double"
				},
				"items": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/UserAttributes"
					}
				}
			},
			"required": [
				"count",
				"items"
			],
			"type": "object"
		},
		"UserCreateRequest": {
			"properties": {
				"email": {
					"type": "string",
					"format": "email",
					"pattern": "^[a-zA-Z0-9_.+-]+\\x40[a-zA-Z0-9-]+\\.[a-zA-Z0-9-.]+$"
				},
				"name": {
					"type": "string",
					"minLength": 1
				},
				"role": {
					"type": "string"
				},
				"password": {
					"type": "string"
				}
			},
			"required": [
				"email",
				"name",
				"role",
				"password"
			],
			"type": "object"
		},
		"UserRequestData": {
			"properties": {
				"email": {
					"type": "string",
					"format": "email",
					"pattern": "^[a-zA-Z0-9_.+-]+\\x40[a-zA-Z0-9-]+\\.[a-zA-Z0-9-.]+$"
				},
				"name": {
					"type": "string",
					"minLength": 1
				},
				"role": {
					"type": "string"
				},
				"password": {
					"type": "string",
					"x-nullable": true
				}
			},
			"required": [
				"email",
				"name",
				"role"
			],
			"type": "object"
		},
		"AuthResponse": {
			"properties": {
				"token": {
					"type": "string"
				}
			},
			"required": [
				"token"
			],
			"type": "object"
		},
		"LoginRequest": {
			"properties": {
				"email": {
					"type": "string",
					"format": "email",
					"pattern": "^[a-zA-Z0-9_.+-]+\\x40[a-zA-Z0-9-]+\\.[a-zA-Z0-9-.]+$"
				},
				"password": {
					"type": "string",
					"minLength": 1
				}
			},
			"required": [
				"email",
				"password"
			],
			"type": "object"
		}
	},
	"info": {
		"title": "tsoa-api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		}
	},
	"paths": {
		"/users/{id}": {
			"get": {
				"operationId": "Get",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/UserAttributes"
						}
					},
					"400": {
						"description": "Bad request"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"tags": [
					"users"
				],
				"security": [
					{
						"jwt": [
							"user:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"format": "double",
						"type": "number"
					}
				]
			},
			"put": {
				"operationId": "Update",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/UserAttributes"
						}
					},
					"400": {
						"description": "Bad request"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"tags": [
					"users"
				],
				"security": [
					{
						"jwt": [
							"user:write"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"format": "double",
						"type": "number",
						"pattern": "^[A-Fa-f\\d]{24}$"
					},
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/UserRequestData"
						}
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "Bad request"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"tags": [
					"users"
				],
				"security": [
					{
						"jwt": [
							"user:write"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"format": "double",
						"type": "number",
						"pattern": "^[A-Fa-f\\d]{24}$"
					}
				]
			}
		},
		"/users": {
			"get": {
				"operationId": "Search",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/PaginationResponseUserAttributes[]"
						}
					},
					"400": {
						"description": "Bad request"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					}
				},
				"tags": [
					"users"
				],
				"security": [
					{
						"jwt": [
							"user:read"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"format": "int32",
						"type": "integer",
						"minimum": 1
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"format": "int32",
						"type": "integer",
						"minimum": 1,
						"maximum": 100
					},
					{
						"in": "query",
						"name": "query",
						"required": false,
						"type": "string"
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"type": "string"
					},
					{
						"in": "query",
						"name": "sortDirection",
						"required": false,
						"type": "string",
						"enum": [
							"ASC",
							"DESC"
						],
						"pattern": "^ASC|DESC$"
					}
				]
			},
			"post": {
				"operationId": "Create",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/UserAttributes"
						}
					},
					"400": {
						"description": "Bad request"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					}
				},
				"tags": [
					"users"
				],
				"security": [
					{
						"jwt": [
							"user:write"
						]
					}
				],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/UserCreateRequest"
						}
					}
				]
			}
		},
		"/auth": {
			"post": {
				"operationId": "Login",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/AuthResponse"
						}
					},
					"400": {
						"description": "Bad request"
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"tags": [
					"auth"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/LoginRequest"
						}
					}
				]
			}
		}
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {
		"jwt": {
			"type": "apiKey",
			"name": "Authorization",
			"in": "header"
		}
	}
}